+++ app/Models/Contact.php
@@ -348,6 +348,16 @@ class Contact extends Model implements CalendarInterface
     }
 
     /**
+     * Defines the has-many relationship with the GiftIdea model
+     *
+     * @return \Illuminate\Database\Eloquent\Relations\HasMany
+     */
+    public function giftIdeas()
+    {
+        return $this->hasMany(GiftIdea::class);
+    }
+
+    /**
      * Get the route key for the model.
      *
      * @return string